backend/requirements.txt
fastapi==0.115.8
fastapi-cli==0.0.7
pydantic==2.10.6
pymongo==4.11.2
python-dotenv==1.2.2
python-jose==3.5.0

